requirements.txt
pandas
numpy
sqlite3
python-dateutil
requests
beautifulsoup4
vader-sentiment
transformers
torch
apscheduler
logging
fastapi
uvicorn
streamlit
yfinance
plotly
streamlit-autorefresh
exchange-calendars
pytest